refactor(ia): new error handling, removed `/tabs` directory
